requirements.txt
flask>=3.1.1
flask-jwt-extended>=4.7.1
flask-migrate>=4.1.0
flask-sqlalchemy>=3.1.1
pydantic>=2.0
python-dotenv>=1.0.0
requests>=2.31.0
